How to use it?
It's simple, you have to get video of your liking to the assets folder, then change path inside html file.
You got this!
I couldn't push this repository to github, but I was struggling if I should delete it, so this is the idea - it will stay on github this time.